Added Adapters / Smart Shutdown
Adapters are a way of describing and generisizing different methods
to handle processing data. The default adapter is an HttpAdapter that
handles any URLs that use http(s) schemas.

An example future adapter could be a File adapter that copies a file
to a path, or an RSS / Video adapter for handling RSS feeds and videos
respectively.

This commit also adds smarter shutdown. In order to create adapters,
an event loop rather than a "request / response" loop was implemented.
This meant that an adapter could send an update message to the Daemon
during the Daemon's shutdown, this caused a bug where a downloaded
file would successfully download, but the `OfflineData` update would
never be realized.

Though the main loop is now slightly uglier, the Daemon now supports
much lower timeouts and can effectively "save work" and properly
shutdown.