Added an RPM Spec file
This is a file that tells rpm build tools how to create a package.
Currently It will only output an x86 package (Someone might want to
cross compile later). It should be noted that this spec file expects
to have monk already compiled before build time. To change this, uncomment
the line in the %build section. The RPM that this generates will install
monk in /usr/bin for a global install. Some low hanging fruit for this
would be modifing %build to check if it has a binary. CICD still has to
be configured to build an RPM artifact.